refactor: move platform into folder with proper isolation
