+++ packages/hoppscotch-common/src/platform/index.ts
@@ -1,0 +1,13 @@
+import { AuthPlatformDef } from "./auth"
+import { UIPlatformDef } from "./ui"
+
+export type PlatformDef = {
+  ui?: UIPlatformDef
+  auth: AuthPlatformDef
+}
+
+export let platform: PlatformDef
+
+export function setPlatformDef(def: PlatformDef) {
+  platform = def
+}
