+++ packages/hoppscotch-common/src/platform/ui.ts
@@ -1,0 +1,8 @@
+import { Ref } from "vue"
+
+export type UIPlatformDef = {
+  appHeader?: {
+    paddingTop?: Ref<string>
+    paddingLeft?: Ref<string>
+  }
+}
